refactor: Move Thumb to it's own component
